Removed catalina.out on restart, and related refactoring.
