+++ puppet/modules/orcid_deployment/files/home/orcid_tomcat/bin/scripts/deployment/restart-tomcat.py
@@ -3,4 +3,5 @@
 import shared
 
 shared.stop_tomcat()
+shared.clean_tomcat()
 shared.start_tomcat()
